+++ src/FedEx/LocationsService/SimpleType/NotificationSeverityType.php
@@ -1,0 +1,20 @@
+<?php
+namespace FedEx\LocationsService\SimpleType;
+
+use FedEx\AbstractSimpleType;
+
+/**
+ * NotificationSeverityType
+ *
+ * @author      Jeremy Dunn <jeremy@jsdunn.info>
+ * @package     PHP FedEx API wrapper
+ * @subpackage  Locations Service
+ */
+class NotificationSeverityType extends AbstractSimpleType
+{
+    const _ERROR = 'ERROR';
+    const _FAILURE = 'FAILURE';
+    const _NOTE = 'NOTE';
+    const _SUCCESS = 'SUCCESS';
+    const _WARNING = 'WARNING';
+}